video-dashboard/requirements.txt
flask>=3.0,<4
flask-socketio>=5.3,<6
simple-websocket>=1.0,<2
python-dotenv>=1.0,<2
requests>=2.31
httpx>=0.27
werkzeug>=3.0
azure-identity>=1.15
